# openapi.yaml
openapi: 3.0.3
...
info:
  title: Ryden REST API
  version: "1.0"
  description: >
    API REST (Jersey/JAX-RS) de Ryden — plataforma P2P de alquiler de autos.
    Reemplaza la capa Spring MVC + JSP. Consumida por una SPA servida como
    estáticos desde Tomcat (sin SSR). Esta especificación es el contrato
    público (DTOs), independiente del modelo de dominio.
servers:
  - url: /webapp
    description: >
      Context path en Tomcat local (WAR desplegado como webapp.war).
      En Jetty dev (`mvn jetty:run`) el context path es `/` — ajustar el prefijo al probar.
  - url: /
    description: Raíz de la API en Jetty dev o despliegue sin context path
...
security:
  - bearerAuth: []
paths:
...
  /cars:
    get:
      tags: [cars]
      summary: Buscar / listar autos (reemplaza /search, home, /my-cars, /admin/cars)
      description: >
        Una sola colección con TODOS los filtros como query params. Las
        distintas pantallas del MVC (home cheapest/recent, búsqueda pública,
        "mis autos", listado admin) colapsan acá variando query params +
        autorización. Paginación por header `Link`. Semántica según query
        (decisión documentada, ver auditoría A14): (1) `scope=admin` → TODO
        el catálogo sin filtro de estado, solo admin; (2) `ownerId` sin
        `scope` → "autos de este usuario": si el llamante es ese usuario o
        admin, ve cualquier estado (o el filtro `status` explícito que
        mande); cualquier otro llamante ve solo `active` y no puede pasar
        `status` (403); (3) sin `ownerId` ni `scope` → browse público,
        siempre `active` únicamente.         Es un filtro de colección legítimo
        (no una operación distinta por rol), pero al ser 3 ramas con
        autorización propia se documenta acá en vez de separarlo en
        3 endpoints distintos.
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - name: q
          in: query
          schema: { type: string }
          description: Texto libre (marca/modelo/descripción).
        - name: ownerId
          in: query
          schema: { type: integer, format: int64 }
          description: Filtra por dueño (sustituye "/my-cars" pasando el propio id).
        - name: category
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: "#/components/schemas/CarType" } }
          description: Repetible (`category=suv&category=sedan`) para filtrar por varias categorías a la vez.
        - name: transmission
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: "#/components/schemas/Transmission" } }
          description: Repetible, igual que `category`.
        - name: powertrain
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: "#/components/schemas/Powertrain" } }
          description: Repetible, igual que `category`.
        - name: priceMin
          in: query
          schema: { type: number }
        - name: priceMax
          in: query
          schema: { type: number }
        - name: rating
          in: query
          style: form
          explode: true
          schema: { type: array, items: { type: number, minimum: 0, maximum: 5 } }
          description: Repetible; cada valor es un piso de calificación mínima (p.ej. `rating=4&rating=5`).
        - name: neighborhoodId
          in: query
          schema: { type: integer, format: int64 }
        - name: from
          in: query
          schema: { type: string, format: date }
          description: Inicio de la ventana de disponibilidad (wall date).
        - name: until
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: "#/components/schemas/CarStatus" } }
          description: >
            Repetible. Filtro por estado. Solo el dueño (vía ownerId propio) o un admin pueden
            filtrar por estados no públicos.
        - name: scope
          in: query
          schema: { type: string, enum: [admin] }
          description: >
            scope=admin lista TODOS los autos del catálogo, cualquier estado, paginados
            (reemplaza la vieja vista /admin/cars). Solo admin (403 en caso contrario).
            Distinto del browse público, que solo expone autos publicados.
        - name: sort
          in: query
          schema: { type: string, enum: [price_asc, price_desc, rating_desc, recent, name] }
      responses:
        "403":
          description: >
            scope=admin sin ser admin, o status explícito en ownerId de otro usuario sin ser admin.
          content:
            application/vnd.paw.error.v1+json:
              schema: { $ref: "#/components/schemas/ErrorDto" }
        "200":
          description: Lista de autos.
          headers:
            Link: { $ref: "#/components/headers/Link-Pagination" }
            X-Total-Count: { $ref: "#/components/headers/X-Total-Count" }
          content:
            application/vnd.paw.car.v1+json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CarDto" }
            application/vnd.paw.car.v1+xml:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CarDto" }
        "204": { description: Sin resultados. }
    post:
      tags: [cars]
      summary: Publicar un auto (reemplaza POST /publish-car)
      description: >
        Crea el Car. Según prerequisitos (identidad+cbu) y si la marca/modelo
        está validada, el server fija el `status` inicial (active / lack_doc) y
        si el modelo quedó pendiente lo indica en la representación. La galería
        y el seguro se suben luego como sub-recursos (PUT /cars/{id}/pictures,
        PUT /cars/{id}/insurance) o en el mismo POST multipart.
      requestBody:
        required: true
        content:
          application/vnd.paw.car.v1+json:
            schema: { $ref: "#/components/schemas/CarCreateDto" }
          multipart/form-data:
            schema:
              type: object
              properties:
                car: { $ref: "#/components/schemas/CarCreateDto" }
                pictures:
                  type: array
                  items: { type: string, format: binary }
                insurance: { type: string, format: binary }
      responses:
        "201":
          description: Auto publicado.
          headers:
            Location: { $ref: "#/components/headers/Location" }
          content:
            application/vnd.paw.car.v1+json:
              schema: { $ref: "#/components/schemas/CarDto" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: No cumple prerequisitos (identidad/cbu) o plate duplicada.
  /cars/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [cars]
      summary: Detalle de un auto
      responses:
        "200":
          description: Auto.
          content:
            application/vnd.paw.car.v1+json:
              schema: { $ref: "#/components/schemas/CarDto" }
            application/vnd.paw.car.v1+xml:
              schema: { $ref: "#/components/schemas/CarDto" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [cars]
      summary: Reemplazar atributos del auto (owner)
      requestBody:
        required: true
        content:
          application/vnd.paw.car.v1+json:
            schema: { $ref: "#/components/schemas/CarDto" }
      responses:
        "200":
          description: Auto reemplazado.
          content:
            application/vnd.paw.car.v1+json:
              schema: { $ref: "#/components/schemas/CarDto" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [cars]
      summary: Transiciones de estado del auto (pausar/despausar/admin-pausar/desactivar)
      description: >
        Absorbe TODAS las "acciones" del MVC sobre el auto:
        - Owner pausa/despausa → { "status": "paused" | "active" } (toggle viejo).
        - Owner desactiva → { "status": "deactivated" } (irreversible). [También
          podría ser DELETE como soft-delete; ver decisión D5.]
        - Admin pausa/despausa → { "status": "admin_paused" | "active" } (solo
          admin; reemplaza /admin/cars/{id}/pause y /resume). Dispara
          cancelación en cascada de reservas y email al owner.
        - Editar precio/min rental days / descripción → atributos sueltos.
        El resource valida transición permitida (405/403/400 según el caso) y
        que el llamante tenga permiso (owner vs admin).
      requestBody:
        required: true
        content:
          application/vnd.paw.car.v1+json:
            schema: { $ref: "#/components/schemas/CarPatchDto" }
      responses:
        "200":
          description: Auto actualizado.
          content:
            application/vnd.paw.car.v1+json:
              schema: { $ref: "#/components/schemas/CarDto" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Transición no permitida desde el estado actual.
    delete:
      tags: [cars]
      summary: Desactivar (soft-delete) el auto (owner) — ver decisión D5
      responses:
        "204": { description: Desactivado. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --- Disponibilidad: sub-recurso fuerte con identidad propia ---
...
  /cars/{id}/pictures:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [cars, files]
      summary: Galería del auto (lista de metadatos hipervinculados)
      responses:
        "200":
          description: Items de galería.
          content:
            application/vnd.paw.picture.v1+json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PictureDto" }
    post:
      tags: [cars, files]
      summary: Agregar una imagen/video a la galería (owner)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                displayOrder: { type: integer }
      responses:
        "201":
          description: Item agregado.
          headers:
            Location: { $ref: "#/components/headers/Location" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /cars/{id}/pictures/{pictureId}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: pictureId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [cars, files]
      summary: Bytes de una imagen/video de la galería (público)
      responses:
        "200": { $ref: "#/components/responses/BinaryImage" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [cars, files]
      summary: Quitar item de la galería (owner)
      responses:
        "204": { description: Quitado. }
        "403": { $ref: "#/components/responses/Forbidden" }

  # --- Seguro del auto (sub-recurso binario débil) ---
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        Authorization: Basic base64(email:secret) a CUALQUIER endpoint.
        El secret depende del contexto:
        - Cuenta verificada: password (BCrypt).
        - Verificación de email (cuenta sin verificar): OTP de 6 dígitos enviado
          por mail tras POST /users. Se consume al autenticar; devuelve JWT.
        - Reset de password olvidado: OTP enviado tras POST /credentials. Otorga
          authority ROLE_PASSWORD_RESET_OTP; se consume al completar PATCH
          /users/{id} con password + passwordConfirm.
        Tras auth válida el server responde con X-Access-Token, X-Refresh-Token
        y Link rel="authenticated-user". No hay /login ni /verify.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Requests normales: Authorization: Bearer <access_token>. Cuando vence
        (401), se manda el Refresh Token (también como Bearer) a cualquier
        endpoint y el server devuelve un par nuevo en headers. No hay /refresh:
        el JWT es autocontenido y el server distingue access de refresh.
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (la metadata de paginación va en el header Link, no en el body).
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 12 }
  headers:
    Location:
      description: URN canónica del recurso recién creado.
      schema: { type: string, format: uri }
    Link-Pagination:
      description: >
        Links de paginación con rel="first" | "prev" | "next" | "last",
        honrando los query params de filtro activos. El cliente los navega
        genéricamente; NO construye URLs a mano.
      schema: { type: string }
      example: '</cars?page=2&category=suv>; rel="next", </cars?page=5&category=suv>; rel="last"'
    X-Total-Count:
      description: Cantidad total de elementos (opcional, complementa Link).
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: 401 — Falta autenticación o token inválido/vencido.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
    Forbidden:
      description: 403 — Autenticado pero sin permiso.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
    NotFound:
      description: 404 — El recurso no existe.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
...
    ValidationError:
      description: 400 — Request mal formado o validación fallida.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ValidationErrorDto" }
    BinaryImage:
      description: Bytes de una imagen (content-type detectado).
      content:
        image/*:
          schema: { type: string, format: binary }
...
  schemas:
    Links:
      type: object
      description: Bloque de hipervínculos generado on-the-fly. Siempre incluye self. NUNCA se persiste.
      additionalProperties: { type: string, format: uri }
      properties:
        self: { type: string, format: uri }
      required: [self]
...
    CarType:
      type: string
      enum: [sedan, hatchback, suv, coupe, convertible, wagon, van, pickup]
    Powertrain:
      type: string
      enum: [gasoline, diesel, electric, hybrid, cng]
    Transmission:
      type: string
      enum: [manual, automatic, semi_automatic]
    CarStatus:
      type: string
      enum: [active, paused, admin_paused, lack_doc, unavailable, deactivated]
    PriceMarketPosition:
      type: string
      enum: [below_market, at_market, above_market]
      description: Consumer browse card badge relative to same brand/model market average.
    CarDto:
      type: object
      properties:
        plate: { type: string }
        year: { type: integer, nullable: true }
        powertrain: { $ref: "#/components/schemas/Powertrain" }
        transmission: { $ref: "#/components/schemas/Transmission" }
        type: { $ref: "#/components/schemas/CarType" }
        status: { $ref: "#/components/schemas/CarStatus" }
        description: { type: string, nullable: true }
        minimumRentalDays: { type: integer }
        ratingAvg: { type: number, nullable: true }
        dayPrice:
          type: number
          nullable: true
          description: Lowest active day price; populated on browse card projections only.
        brandName: { type: string }
        modelName: { type: string }
        modelValidated: { type: boolean }
        hasInsurance: { type: boolean }
        priceMarketPositionModifier:
          allOf:
            - $ref: "#/components/schemas/PriceMarketPosition"
          nullable: true
          description: Consumer browse/favorites cards when comparable market data exists.
        marketAveragePrice:
          type: number
          nullable: true
          description: Same brand/model market average day price backing the badge.
        marketSampleCount:
          type: integer
          format: int64
          nullable: true
          description: Number of comparable active listings used for the badge.
        createdAt: { type: string, format: date-time }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self, owner, model, brand, pictures, availabilities, insurance, reviews.
    CarCreateDto:
      type: object
      required: [plate, transmission, powertrain]
      properties:
        plate: { type: string }
        year: { type: integer }
        powertrain: { $ref: "#/components/schemas/Powertrain" }
        transmission: { $ref: "#/components/schemas/Transmission" }
        description: { type: string }
        minimumRentalDays: { type: integer, default: 1 }
        # Catálogo: por nombre o por URN de modelo existente.
        brandName: { type: string }
        modelName: { type: string }
        type: { $ref: "#/components/schemas/CarType" }
        modelUri: { type: string, format: uri, description: Alternativa a brand/model name si el modelo ya existe (se navega su link). }
    CarPatchDto:
      type: object
      description: Transiciones de estado y ediciones parciales del auto.
      properties:
        status: { $ref: "#/components/schemas/CarStatus" }
        description: { type: string }
        minimumRentalDays: { type: integer }
...
    PictureDto:
      type: object
      properties:
        displayOrder: { type: integer }
        kind: { type: string, enum: [image, video] }
        contentType: { type: string }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self (bytes), car.
...
    ErrorDto:
      type: object
      properties:
        status: { type: integer }
        code: { type: string }
        message: { type: string }
    ValidationErrorDto:
      type: object
      properties:
        status: { type: integer, example: 400 }
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }